Destroying keys in key store v2
Add a new low-level key store API method DestroyKey() which destroys
keys. That is, it removes key data and sets the state to "destroyed",
but it does not remove the key entirely from the key store. Attempt
to access the key will result in a "key destroyed" error rather than
"key missing".

Later we might introduce some "delete key" or "remove key" method to be
able to purge destroyed keys from the store completely. For now, it's
important to keep the keys in order to keep the sequential numbering.