Teach acra-key-tool to destroy keys
Add new option "--destroy_key" which commands the tool to destroy the
specified key for given client ID. At the moment we need to be able to
destroy only transport key pairs so that's what we support for now.

Note that "--read_key" and "--destroy_key" options are mutually
exclusive. You can do one operation at a time.