Zeroize keys after printing
Make sure we wipe unencrypted keys from memory as soon as we no longer
need them. This may be unnecessary -- because the process is about to
exit anyway -- but it's still good hygiene.

Note that we use a separate function, not main(), because deferred
function calls are performed on function return, not scope exit.