Test key removal in key store v2 (#371)
* Separate key stores: TestKeyNonExistence

Instead of using global key store for everything, use two separate
temporary directories for keys: one for acra-server and one for
acra-connector that we're launching here. Initially the key stores
are identical but we'll be removing keys from them for testing.

Currently we support only key store v1 and the key store is simply
copied (as there is no key export utility key). Later we'll update
this test case to replicate actual key exchage more accurately.

We also update the create_client_keypair() function with an optional
argument to override the key directory it acts upon. Key store version
is still global (KEYSTORE_VERSION).

* Separate key stores: TestKeyStorageClearing

Just like with TestKeyNonExistence, keep test key store separate and
distinct from the global key store. Reuse the existing code.

* Rename acra-read-key => acra-key-tool

This utility is going to be out "multitool" for key manipulation.
Initially it was thought to be useful only for testing but it seems
that it will be a good point to introduce key exporting functionality
and it may also replace acra-keymaker in the future.

However, that's in the future. Right now, rename the tool to prepare it
for extensions, rename the command-line option to be more specific, and
move key reading implementation into a separate function.

* Destroying keys in key store v2

Add a new low-level key store API method DestroyKey() which destroys
keys. That is, it removes key data and sets the state to "destroyed",
but it does not remove the key entirely from the key store. Attempt
to access the key will result in a "key destroyed" error rather than
"key missing".

Later we might introduce some "delete key" or "remove key" method to be
able to purge destroyed keys from the store completely. For now, it's
important to keep the keys in order to keep the sequential numbering.

* Destroying keys by purpose

Add a set of new high-level key store API methods to destroy keys by
their purpose. For now we only need to handle transport key pairs.
Destroying the keys requires the same credentials as creating them
so the methods are added to the TransportKeyCreation interface used by
acra-keymaker but not available to general services.

Key store v1 destroys the keys by removing the key file. Key store v2
uses newly added destruction capabilities for the same effect. There
is a slight difference in implementation: key store v2 retains the
metadata of the key while v1 removes the file completely. Key store v1
has way too many places that depend on the file existence or absence
so we need to remove it rather than truncate it (remove only content).
The concept of "destroyed" key is supported only since key store v2.

* Teach acra-key-tool to destroy keys

Add new option "--destroy_key" which commands the tool to destroy the
specified key for given client ID. At the moment we need to be able to
destroy only transport key pairs so that's what we support for now.

Note that "--read_key" and "--destroy_key" options are mutually
exclusive. You can do one operation at a time.

* Use acra-key-tool to destroy keys in tests

Update the destroy_key() implementation to support key store v2 by using
the common tool instead of directly removing files from the filesystem.

* Enable key removal tests for key store v2

Finally, we are able to unskip those tests and verify key removal path
during regular CI runs.

* Zeroize keys after printing

Make sure we wipe unencrypted keys from memory as soon as we no longer
need them. This may be unnecessary -- because the process is about to
exit anyway -- but it's still good hygiene.

Note that we use a separate function, not main(), because deferred
function calls are performed on function return, not scope exit.