Leave flag.Args() intact
Config file parsing resets command-line arguments to whatever is in the
configuration file (or to an empty string if there is no configuration
file). This makes flag.Args() useless.

Make sure that flag.Args() returns remaining parameters specified on
the command-line after parsing all options. This enables support of
subcommands.