Improve --help for arca-keys
Make the common pretty-printing utility reusable with other flag sets
and improve "Usage" implementation for acra-keys to print out the entire
command and flag sets.

It is also possible to get help for a particular command like this:

    acra-keys read --help

Global help is still not ideal, but much better for subcommands now.