Handle rotated keys in acra-rollback
Teach acra-rollback to handle rotated storage keys. That is, use
DecryptRotatedAcrastruct() that takes into account historical keys
and is able to decrypt data even after keys have been rotated without
reencrypting the old data.

While we're here, don't forget to wipe the keys from memory once we no
longer need them.

Verify that this actually works. For that we need to teach
create_client_keypair() to rotate only storage keys to not disturb
connections we open before the test.

Also, make sure the tables are created before the test starts. It's
helpful if only this test is executed. Before we relied on previous
tests to create the test table.