Handle rotated keys in acra-rotate
Similarly, teach acra-rotate to use historical keys for decryption while
using the latest key to encrypt the data.

We already have tests that verify acra-rotate operation. They are not
simple and I'd like to avoid massive duplication. Pull in a library for
data-driven testing from PyPI, this should allow to insert the necessary
key rotation call with minimal duplication.

Unfortunately, there is no way to rotate zone keys with utilities right
now (acra-addzone adds a new zone, it does not touch previous one). We
do not add tests for that, but it should be possible to implement in
a similar way once zone key rotation is implemented.

TestAcraRotate test suite inherits from TestAcraRotateWithZone,
including the "testDatabaseRotation" method. It overrides the method,
but this does not work as expected with DDT's decorators. Skip the
inherited test to avoid failures (it does not work without zones).
It's ugly, but it works.