Handle rotate keys in acra-server
Finally, teach acra-server to decrypt historical data. And wipe the keys
once we no longer need them. This is important since acra-server is long
running service so we should avoid keeping the keys in memory for more
than strictly necessary.

Add a new test to verify key rotation. Like with acra-rotate, we check
only the default zone mode because it is not possible to rotate zone
keys. However, it should work if the keys are actually rotated.