Make interface changes additive
Only add new methods to interfaces instead of changing defitions of
existing GetPrivateKey() and ReadSymmetricKey(). This allows existing
code outside of "decryptor" package to retain its behavior of using
only the latest key for decryption.

However, this change will still break Acra EE code that extends
Decryptor from Acra CE. This is intentional, it makes it easier to track
places that need to be updated in Acra EE.