Key management multi-tool: "acra-keys" (#374)
* Rename acra-key-tool => acra-keys

The collective will has decided that this will be a better name for
a tool that's going to be used for key management. Eventually
acra-keymanager will be made a part of it too.

* Move acra-keys logic into package

This way it will be possible to reuse this code in Acra EE without much
copypasta to handle keys supported by Acra CE.

* Sub-command based CLI for acra-keys

Replace "--read_key" and "--destroy_key" options with sub-commands
"read" and "destroy". Now the keys may be specified like this:

    acra-keys --client_id Alice read storage-public

* Leave flag.Args() intact

Config file parsing resets command-line arguments to whatever is in the
configuration file (or to an empty string if there is no configuration
file). This makes flag.Args() useless.

Make sure that flag.Args() returns remaining parameters specified on
the command-line after parsing all options. This enables support of
subcommands.

* Improve --help for arca-keys

Make the common pretty-printing utility reusable with other flag sets
and improve "Usage" implementation for acra-keys to print out the entire
command and flag sets.

It is also possible to get help for a particular command like this:

    acra-keys read --help

Global help is still not ideal, but much better for subcommands now.

* Reduce duplication of key option checks

Move all command-line validation into a CheckForKeyKind() method of
CommandLineParams to keep it in place.