Support storage key rotation (#375)
* Handle rotated keys in acra-rollback

Teach acra-rollback to handle rotated storage keys. That is, use
DecryptRotatedAcrastruct() that takes into account historical keys
and is able to decrypt data even after keys have been rotated without
reencrypting the old data.

While we're here, don't forget to wipe the keys from memory once we no
longer need them.

Verify that this actually works. For that we need to teach
create_client_keypair() to rotate only storage keys to not disturb
connections we open before the test.

Also, make sure the tables are created before the test starts. It's
helpful if only this test is executed. Before we relied on previous
tests to create the test table.

* Handle rotated keys in acra-rotate

Similarly, teach acra-rotate to use historical keys for decryption while
using the latest key to encrypt the data.

We already have tests that verify acra-rotate operation. They are not
simple and I'd like to avoid massive duplication. Pull in a library for
data-driven testing from PyPI, this should allow to insert the necessary
key rotation call with minimal duplication.

Unfortunately, there is no way to rotate zone keys with utilities right
now (acra-addzone adds a new zone, it does not touch previous one). We
do not add tests for that, but it should be possible to implement in
a similar way once zone key rotation is implemented.

TestAcraRotate test suite inherits from TestAcraRotateWithZone,
including the "testDatabaseRotation" method. It overrides the method,
but this does not work as expected with DDT's decorators. Skip the
inherited test to avoid failures (it does not work without zones).
It's ugly, but it works.

* Handle rotate keys in acra-server

Finally, teach acra-server to decrypt historical data. And wipe the keys
once we no longer need them. This is important since acra-server is long
running service so we should avoid keeping the keys in memory for more
than strictly necessary.

Add a new test to verify key rotation. Like with acra-rotate, we check
only the default zone mode because it is not possible to rotate zone
keys. However, it should work if the keys are actually rotated.

* Consistently use ZeroizePrivateKey

* Test rotation with old and new keys

Make sure that we really can handle both old and new keys by adding some
data before and after rotation. We need to query the key again to read
the latest value from the key store. Decryption should transparently
handle both cases and return plaintext data as is.

* Use multiple keys in Decryptor internals

Aside from using DecryptRotatedAcrastruct() Decryptor also has a bunch
of places which decrypt AcraStructs in a streaming fashion. They extract
symmetric key, decrypt it manually, then proceed with the data.

In particular, MySQL always uses this code path.

Teach Decryptor to use multiple keys by default too. Update all relevant
interfaces and locations to iterate over a list of private keys rather
than accept only one.

Note that poison records still use only one key.

* Introduce ZeroisePrivateKeys utility

There are may places where we need to wipe a slice of private keys.
Introduce a utility to do this.

Take care with defers, they need to be used *after* private key slice
assignment.

* Make interface changes additive

Only add new methods to interfaces instead of changing defitions of
existing GetPrivateKey() and ReadSymmetricKey(). This allows existing
code outside of "decryptor" package to retain its behavior of using
only the latest key for decryption.

However, this change will still break Acra EE code that extends
Decryptor from Acra CE. This is intentional, it makes it easier to track
places that need to be updated in Acra EE.