Store standalone public keys as key pairs
Remove a dedicated key format for standalone public keys and store them
as "key pairs", just without a private key.

It was a nice early idea but ultimately it is not convenient because
PublicKey() getter implementation will have to guess whether it needs
to retrieve a public key of a public key or a public key of a key pair.
Since the data format inside is the same -- Themis asymmetric keys --
it does not make much sense to have a separate format for that.

Make the ASN.1 value reserved, remove Go constants, and update the code
to allow and handle lone public keys in a predictable way. That is, the
users are allowed to store key pairs without private keys but they will
get a relevant error if they try to retrieve the private part.

Standalone public keys are an important feature for key exchange. They
will be exported and moved to different machines which need to interact
with the one that generated a key pair.