ASN.1 data definitions for exported keys
We are going to store exported keys in encrypted and signed form. We
need a signature to verify that the data has not been tampered with.
Data is also stored encrypted to prevent unauthorized access to key
material, obviously.

Keys are normally stored encrypted too, but we do not know where they
will be stored when exporting so we can't encrypt then with a correct
key. Therefore instead of encrypting each key individually, we encrypt
the entire package. This is more efficient and it also hides the
information from a casual explorer with access to backups.

So the idea is to put a set of exported KeyRings into EncryptedKeys
structure, then serialize it, encrypt it, and put that as octet string
into a SignedContainer.