Key export and import API
Introduce a pair of API methods on KeyStores to export and import sets
of key rings. Specified key rings can be exported with ExportKeyRings().
They will be retrieved from the key store, decrypted using the key store
key, then encrypted, packed, and signed with the provided cryptosuite
(normally, with a newly generated "export" set of keys).

Export key blob can then be transferred to another machine for import.
Import is performed with ImportKeyRings() which basically reverses the
process: verify the signature, unpack and decrypt key rings, then suck
them into the key store, encrypting them with the new key store keys.

Note that importing is more complicated because there are more failures
possible. The key rings might be already present in the key store (in
full or partially), and the importing process might fail midway. The
process may be controlled with a delegate.

Also note that we handle sensitive data -- decrypted key material. We
take great care to erase it from memory as fast as possible, when we no
longer need it. This includes the failure path too where we can have
some intermediate results that cannot be used due to the failure.

This commit implements only exporting path and some high-level parts of
the import process. Details will be fleshed out in the next commit.