Implement key import
Add all the missing stuff to implement key import in the backend as
well. Note that we need to open a new key ring if it does not exist
yet before importing new data into it. The import process itself
uses a new type of transaction that sets the key material --
basically, it's always overwriting the key data with imported data.

Test for all import modes are a little bit verbose, but they do
their job well.