Key store v2 export/import API (#376)
* Store standalone public keys as key pairs

Remove a dedicated key format for standalone public keys and store them
as "key pairs", just without a private key.

It was a nice early idea but ultimately it is not convenient because
PublicKey() getter implementation will have to guess whether it needs
to retrieve a public key of a public key or a public key of a key pair.
Since the data format inside is the same -- Themis asymmetric keys --
it does not make much sense to have a separate format for that.

Make the ASN.1 value reserved, remove Go constants, and update the code
to allow and handle lone public keys in a predictable way. That is, the
users are allowed to store key pairs without private keys but they will
get a relevant error if they try to retrieve the private part.

Standalone public keys are an important feature for key exchange. They
will be exported and moved to different machines which need to interact
with the one that generated a key pair.

* ASN.1 data definitions for exported keys

We are going to store exported keys in encrypted and signed form. We
need a signature to verify that the data has not been tampered with.
Data is also stored encrypted to prevent unauthorized access to key
material, obviously.

Keys are normally stored encrypted too, but we do not know where they
will be stored when exporting so we can't encrypt then with a correct
key. Therefore instead of encrypting each key individually, we encrypt
the entire package. This is more efficient and it also hides the
information from a casual explorer with access to backups.

So the idea is to put a set of exported KeyRings into EncryptedKeys
structure, then serialize it, encrypt it, and put that as octet string
into a SignedContainer.

* Key export and import API

Introduce a pair of API methods on KeyStores to export and import sets
of key rings. Specified key rings can be exported with ExportKeyRings().
They will be retrieved from the key store, decrypted using the key store
key, then encrypted, packed, and signed with the provided cryptosuite
(normally, with a newly generated "export" set of keys).

Export key blob can then be transferred to another machine for import.
Import is performed with ImportKeyRings() which basically reverses the
process: verify the signature, unpack and decrypt key rings, then suck
them into the key store, encrypting them with the new key store keys.

Note that importing is more complicated because there are more failures
possible. The key rings might be already present in the key store (in
full or partially), and the importing process might fail midway. The
process may be controlled with a delegate.

Also note that we handle sensitive data -- decrypted key material. We
take great care to erase it from memory as fast as possible, when we no
longer need it. This includes the failure path too where we can have
some intermediate results that cannot be used due to the failure.

This commit implements only exporting path and some high-level parts of
the import process. Details will be fleshed out in the next commit.

* Use interface for delegate

It is possible to pass "nil" for delegate if you're not really
interested in controlling the import details, but if you are interested
then you need to implement all methods of the delegate.

By default any issue is considered a fatal error and the import process
is aborted.

* Implement key import

Add all the missing stuff to implement key import in the backend as
well. Note that we need to open a new key ring if it does not exist
yet before importing new data into it. The import process itself
uses a new type of transaction that sets the key material --
basically, it's always overwriting the key data with imported data.

Test for all import modes are a little bit verbose, but they do
their job well.

* Drop out-of-order rollback checks

Do not perform OOO checks since they are mostly redundant and are not
tested well.