Command-line API for key listing
Introduce "list" subcommand for "acra-keys" utility. It will be used to
enumerate all keys present in the key store, making it easier to decide
which keys the user needs to export.

"list" command is not particularly irreplaceable for filesystem-based
key stores, but it will certainly come handy for remote key stores.

Since we'd like to limit the effort, "list" subcommand is supported only
for keystore v2 in Acra CE right now.