High-level key listing API
Now we translate raw paths into purposes and IDs as defined by our
high-level server/connector/translator key stores. We have to parse
the paths and determine expected purpose of the key from it.

Purpose constants are human-readable but are exported for programming
purposes if necessary. This will probably be useful for key export.