"acra-keys export" subcommand
Add a new subcommand "export" which exports the selected keys. It is
possible to export either an enumerated subset of keys by their IDs
(use "acra-keys list" to query the IDs), or export everything using
the "--all" option.

The keys are encrypted and signed with a freshly generated set of keys
which are provided along with the encrypted exported data. Both pieces
of information are saved into files specified by "--data" and "--keys"
options. Initially they are stored alongside but the users should store
and transmit those two files separately. Encryption keys are provided
in plaintext so certain care is expected for them.

Only key store v2 supports exporting keys. It's not trivial to strap up
the exporting code to key store v1 so we not provide it. The users can
migrate v1 -> v2 (it's free and has no side effects). They can also
request commercial support for Acra EE.