"acra-keys list" subcommand (#377)
* Command-line API for key listing

Introduce "list" subcommand for "acra-keys" utility. It will be used to
enumerate all keys present in the key store, making it easier to decide
which keys the user needs to export.

"list" command is not particularly irreplaceable for filesystem-based
key stores, but it will certainly come handy for remote key stores.

Since we'd like to limit the effort, "list" subcommand is supported only
for keystore v2 in Acra CE right now.

* Low-level key listing API

Implement key listing for low-level filesystem based key store which is
used to store high-level data. At this level we work with "key purposes"
which are basically paths.

New backend method "ListAll" enumerates all available key ring paths
which are then returned by "ListKeyRings" after some massaging (since
filesystem paths have certain internal suffixes for disambiguation).

We always return paths in lexicographic order since that's what
filepath.Walk() does.

* High-level key listing API

Now we translate raw paths into purposes and IDs as defined by our
high-level server/connector/translator key stores. We have to parse
the paths and determine expected purpose of the key from it.

Purpose constants are human-readable but are exported for programming
purposes if necessary. This will probably be useful for key export.

* Printing keys as table

The default output mode for "acra-keys list" is a human-readable table
with key descriptions and IDs. It's nothing fancy but at least it looks
somewhat readable and usable as a general overview.

* Printing keys as JSON

For our machine users we provide an alternate output method: JSON. You
can specify "--json" option and get the key list in machine-readable
format which is easier to parse.

* Make tabular output machine-readable too

Tweak table formatting to be somewhat machine-readable as well. This
makes it possible to parse "acra-keys list" output on machines with
bare minimum like cut & sed or awk, using | as record separator.

* Tests for key list formatting

Verify that we format the keys in expected way. Check tabular output.
Check that we can parse the JSON we output.

* Comment keyring path parsing code

Provide an example of paths that we parse here so that the code is
easier to understand.