"acra-keys import" subcommand
Add a new subcommand "import" which decrypts and imports keys exported
by the "export" subcommand. It uses the same "--data" and "--keys"
options to accept input.

Currently we do not handle import conflicts in any particular way, using
the default action: abort if the imported key ring already exists in the
key store. This is sufficient to transfer keys during the initial setup
and to restore a backup from scratch, but not sufficient to update
rotated keys or roll back existing system to a backup. We will improve
this later with more options.