Export only public keys by default
"acra-keys export" is intended to be used primarily for transferring
public keys between servers running various Acra components. For this
use case private keys must never leave the machine which generated them.
Make sure that we do not export private keys by default. In order to
export private keys (e.g., for backup) the user has to explicitly opt-in
by specifying the "--private_keys" option. Without that option the
private keys will not be decrypted and will not be included into the
exported data set.

Specifics of the export are controlled by a new parameter: ExportMode.
Right now it's binary (public-only or full) but later we might want to
extend it and add more options, so this parameter is a bitmask.

Note that the "private keys" concept includes symmetric keys as well
which are intended for the server-only use and should not be shared.