Skip 'empty' exported keys
If the user did not request private key export but specified a key which
has only private data (e.g., a symmetric key) then instead of exporting
an empty key ring (with full metadata but without actual key material)
skip that key ring entirely.

If we do not skip such key rings, the resulting exported bundle cannot
be imported: we do not allow to import keys without any data which are
not destroyed.

Normally this situation can happen if the user intends to export --all
public keys at the same time to simplify transfer. Instead of making the
user manually spell out all the keys except for the internal symmetric
ones, allow them to do the export easily.

It is not ideal -- from architectural standpoint -- to report skipped
keys to logs directly from the interal implementation details, but it's
so much simpler than a proper notification mechanism which would be an
overkill for this specific use case. So we're rolling with it for now.