Integration test for key exchange sequence
Replace the existing crude code for key exchange with a proper one.
Instead of simply copying the entire key store, do as users would do
in real-world usage: generate the keys for AcraServer and AcraConnector
separately, in their respective key stores, and then exchange only the
public keys.

For key store v1 we have to copy the files manually. acra-keys does not
support export for key store v1. This copying is what the administrators
are expected to do right now to exchange the keys.

For key store v2 (and later) we can use acra-keys to export and then
import appropriate keys. For now, we support only key store v2 so we can
hardcode key IDs there. Nevertheless, this verifies that acra-keys is
capable of transferring keys, and that AcraServer and AcraConnector work
after performing the transfer with acra-keys.