acra-keys export & import subcommands (#378)
* "acra-keys export" subcommand

Add a new subcommand "export" which exports the selected keys. It is
possible to export either an enumerated subset of keys by their IDs
(use "acra-keys list" to query the IDs), or export everything using
the "--all" option.

The keys are encrypted and signed with a freshly generated set of keys
which are provided along with the encrypted exported data. Both pieces
of information are saved into files specified by "--data" and "--keys"
options. Initially they are stored alongside but the users should store
and transmit those two files separately. Encryption keys are provided
in plaintext so certain care is expected for them.

Only key store v2 supports exporting keys. It's not trivial to strap up
the exporting code to key store v1 so we not provide it. The users can
migrate v1 -> v2 (it's free and has no side effects). They can also
request commercial support for Acra EE.

* "acra-keys import" subcommand

Add a new subcommand "import" which decrypts and imports keys exported
by the "export" subcommand. It uses the same "--data" and "--keys"
options to accept input.

Currently we do not handle import conflicts in any particular way, using
the default action: abort if the imported key ring already exists in the
key store. This is sufficient to transfer keys during the initial setup
and to restore a backup from scratch, but not sufficient to update
rotated keys or roll back existing system to a backup. We will improve
this later with more options.

* Print a list of imported key rings

Instead of doing a silent import, confirm the list of affected key
rings. For that we need to have the importing code return the list
of key ring IDs that are affected, and then describe them in the
manner used by "acra-keys list".

* Rename command-line options

Use more explicit option names which we ended up with after discussion.

--data  ==>  --key_bundle_file
--keys  ==>  --key_bundle_secret