Export only public keys by default and test it (#379)
* Export only public keys by default

"acra-keys export" is intended to be used primarily for transferring
public keys between servers running various Acra components. For this
use case private keys must never leave the machine which generated them.
Make sure that we do not export private keys by default. In order to
export private keys (e.g., for backup) the user has to explicitly opt-in
by specifying the "--private_keys" option. Without that option the
private keys will not be decrypted and will not be included into the
exported data set.

Specifics of the export are controlled by a new parameter: ExportMode.
Right now it's binary (public-only or full) but later we might want to
extend it and add more options, so this parameter is a bitmask.

Note that the "private keys" concept includes symmetric keys as well
which are intended for the server-only use and should not be shared.

* Skip 'empty' exported keys

If the user did not request private key export but specified a key which
has only private data (e.g., a symmetric key) then instead of exporting
an empty key ring (with full metadata but without actual key material)
skip that key ring entirely.

If we do not skip such key rings, the resulting exported bundle cannot
be imported: we do not allow to import keys without any data which are
not destroyed.

Normally this situation can happen if the user intends to export --all
public keys at the same time to simplify transfer. Instead of making the
user manually spell out all the keys except for the internal symmetric
ones, allow them to do the export easily.

It is not ideal -- from architectural standpoint -- to report skipped
keys to logs directly from the interal implementation details, but it's
so much simpler than a proper notification mechanism which would be an
overkill for this specific use case. So we're rolling with it for now.

* Unit tests for public-only export

Add a test for implementation details to make sure that private part
of the keypair is not included into the exported data, and that the
symmetric key is skipped completely.

* Integration test for key exchange sequence

Replace the existing crude code for key exchange with a proper one.
Instead of simply copying the entire key store, do as users would do
in real-world usage: generate the keys for AcraServer and AcraConnector
separately, in their respective key stores, and then exchange only the
public keys.

For key store v1 we have to copy the files manually. acra-keys does not
support export for key store v1. This copying is what the administrators
are expected to do right now to exchange the keys.

For key store v2 (and later) we can use acra-keys to export and then
import appropriate keys. For now, we support only key store v2 so we can
hardcode key IDs there. Nevertheless, this verifies that acra-keys is
capable of transferring keys, and that AcraServer and AcraConnector work
after performing the transfer with acra-keys.