Improve reusability of filesystem.KeyExport
The main difference between Acra CE and Acra EE in key exporting is the
list of keys that need to be exported and the method by which they are
exported. Acra EE will need to extend and/or override these methods.

The exporting approach can be readily overridden in ExportPublicKey()
methods, etc. That will be easy to do. However, the key list is
different. Current Acra CE code does not expose the key list itself,
instead it provides the EnumerateExportedKeys() method which queries
the list and processes it using the classifier to determine which key
paths correspond to which keys.

The classification process itself is the same for both Acra CE and EE,
but in the current form it cannot be reused in Acra EE: it is too
coupled with path enumertion. Acre CE supports only raw filesystem, but
Acra EE supports more options and cannot use ioutil.ReadDir() directly.

Introduce a new interface for enumerating the key paths alone, and move
the existing EnumerateExportedKeys() methods into free functions. This
way Acra EE can provide its own path enumeration, its own extended
classifier, but reuse the rest of the export workflow.