Extract key store v1 check into a function
Introduce a new function for key store v1 -- IsKeyDirectory() -- similar
to the one we have for key store v2. It can be used to check whether a
directory looks like a key store. This is a conservative check, and it's
pretty loose for key store v1. However, it still does a good enough job
for acra-keymaker to determine whether it needs to ask the user to set
the key store version explicitly or not.