Swap "-d" and "-v" meaning for consistency
Acra utilities typically use "-v" key to enable slightly more output
(but still useful for normal users) while "-d" is used by developers
for debugging and outputs information not really useful to the users.

Normally, the "-v" maps to Info level, "-d" maps to Debug level, and the
Trace level cannot be enabled by command-line switches. However,
acra-migrate-keys is not a service and it uses Info level liberally for
regular status information.

Keep the log level idea but adjust it for acra-migrate-keys.