Improve acra-migrate-key reusability (#386)
* Common acra-migrate-keys package

Move command-line processing into a shared "migratekeys" package which
could be reused in Acra EE to provide the same command-line interface
as Acra CE without a lot of copypasta.

We'll stil need to use copy-paste inheritance for key store
initialization and the migration process, but that's a necessary evil
because the key store instances are different. Common command-line
parameters are the same, on the other hand.

* Improve reusability of filesystem.KeyExport

The main difference between Acra CE and Acra EE in key exporting is the
list of keys that need to be exported and the method by which they are
exported. Acra EE will need to extend and/or override these methods.

The exporting approach can be readily overridden in ExportPublicKey()
methods, etc. That will be easy to do. However, the key list is
different. Current Acra CE code does not expose the key list itself,
instead it provides the EnumerateExportedKeys() method which queries
the list and processes it using the classifier to determine which key
paths correspond to which keys.

The classification process itself is the same for both Acra CE and EE,
but in the current form it cannot be reused in Acra EE: it is too
coupled with path enumertion. Acre CE supports only raw filesystem, but
Acra EE supports more options and cannot use ioutil.ReadDir() directly.

Introduce a new interface for enumerating the key paths alone, and move
the existing EnumerateExportedKeys() methods into free functions. This
way Acra EE can provide its own path enumeration, its own extended
classifier, but reuse the rest of the export workflow.

* Make key store format required

With new approach, the user is required to explicitly specify the source
and destination key store versions for conversion. The --src_keystore
and --dst_keystore arguments are now required. Print warnings and fail
argument parsing if they are not specified.

* Move argument parsing to common module

It's easier to just do all argument parsing in the common module
altogether instead of leaving it for the tool code.

* Refresh configuration defaults

* Swap "-d" and "-v" meaning for consistency

Acra utilities typically use "-v" key to enable slightly more output
(but still useful for normal users) while "-d" is used by developers
for debugging and outputs information not really useful to the users.

Normally, the "-v" maps to Info level, "-d" maps to Debug level, and the
Trace level cannot be enabled by command-line switches. However,
acra-migrate-keys is not a service and it uses Info level liberally for
regular status information.

Keep the log level idea but adjust it for acra-migrate-keys.