Require explicit key store version (#387)
* Remove --keystore parameter from all tools

Now the key store format is expected to be automatically detected from
the key store data. There is no need for the user to specify or override
it explicitly. Thus, there is no need for the "--keystore" parameter in
most tools. Remove it and refresh default configuration files.

* Make --keystore required for acra-keymaker

In contrast, acra-keymaker now requires the key store version to be
specified explicitly when initializing a new key store. This parameter
is required only during the first run. After that acra-keymaker can
detect the key store version just like all other tools.

* Extract key store v1 check into a function

Introduce a new function for key store v1 -- IsKeyDirectory() -- similar
to the one we have for key store v2. It can be used to check whether a
directory looks like a key store. This is a conservative check, and it's
pretty loose for key store v1. However, it still does a good enough job
for acra-keymaker to determine whether it needs to ask the user to set
the key store version explicitly or not.

* Update default configuration files

* Use --keystore only with acra-keymaker in tests