Improve acra-keys extensibility in Acra EE (#388)
* Use interfaces in acra-keys/keys module

Replace concrete types with interfaces of v2 API so that Acra EE is able
to use import/export functions with its own extended key store.

* Make DescribeKeyRing overridable

This is method is used to describe the keys in the key store for humans
("acra-keys list" uses it, for example). However, in its current form it
is not possible for Acra EE to extend it, providing description for new
keys used only by Acra EE.

Introduce a DescribeKeyRing interface for key store v2 which handles
description on an individual key ring. Acra EE will override this method
to extend it with additonal keys.

Move the main function on a list of keys into a free function which just
calls the abovementioned single-ring method. We could have left it as a
method, but Go does not have default implementations and its bothersome:
easier to have a free function to call from Acra EE.